Divide the implementation in different meaningful files
